small UI updates after feedback
and typing fixes
