+++ public/app/features/alerting/unified/types/contact-points.ts
@@ -1,0 +1,13 @@
+import { IconName } from '@grafana/ui';
+
+export const INTEGRATION_ICONS: Record<string, IconName> = {
+  discord: 'discord',
+  email: 'envelope',
+  googlechat: 'google-hangouts-alt',
+  hipchat: 'hipchat',
+  line: 'line',
+  pagerduty: 'pagerduty',
+  slack: 'slack',
+  teams: 'microsoft',
+  telegram: 'telegram-alt',
+};
